refactor: image 패키지 분리
