+++ matzip-app-external-api/src/main/java/com/woowacourse/matzip/application/image/ImageDeleter.java
@@ -1,4 +1,4 @@
-package com.woowacourse.matzip.application;
+package com.woowacourse.matzip.application.image;
 
 import com.woowacourse.matzip.domain.image.UnusedImage;
 import org.springframework.stereotype.Component;
