+++ matzip-app-external-api/src/main/java/com/woowacourse/matzip/application/image/ImageEventListener.java
@@ -1,4 +1,4 @@
-package com.woowacourse.matzip.application;
+package com.woowacourse.matzip.application.image;
 
 import com.woowacourse.matzip.domain.review.ReviewCreatedEvent;
 import com.woowacourse.matzip.domain.review.ReviewDeletedEvent;
